+++ src/mibig_input_script/aux/parse_arguments.py
@@ -1,3 +1,9 @@
+"""
+Parse command line arguments.
+
+Parses command line arguments for `mibig-input-script.py`.
+"""
+
 import argparse
 from pathlib import Path
 from typing import List
@@ -6,7 +12,7 @@ from mibig_input_script.aux.read_functions import get_curators
 
 
 def add_parser_args(VERSION: str, ROOT: Path) -> argparse.ArgumentParser:
-    """Add arguments to argparse object
+    """Add arguments to argparse object.
 
     Parameters:
         VERSION: str indicating the version
@@ -14,24 +20,13 @@ def add_parser_args(VERSION: str, ROOT: Path) -> argparse.ArgumentParser:
 
     Returns:
         `argparse.ArgumentParser` object
-
-    Notes:
-        Additional command line parameters can be easily specified
-            by adding arguments below.
-
-        Technical:
-        `choices` allows to restrict options
-        `action=store_true` returns True if flag is set
-        `nargs` limits n of args to max 1
     """
     curators = get_curators(ROOT)
 
-    parser = argparse.ArgumentParser(
-        description=f"""
-            mibig-input-script v{VERSION}:
-            Creation and manipulation of MIBiG entries
-        """
+    descr = (
+        f"mibig-input-script v{VERSION}: " "Creation and manipulation of MIBiG entries"
     )
+    parser = argparse.ArgumentParser(description=descr)
 
     parser.add_argument(
         "curator",
@@ -50,30 +45,11 @@ def add_parser_args(VERSION: str, ROOT: Path) -> argparse.ArgumentParser:
         type=str,
         help="Manipulate an existing entry by specifying its MIBiG accession number.",
     )
-
-    # ~ parser.add_argument(
-    # ~ "--gene_ann",
-    # ~ action="store_true",
-    # ~ help="TBA Add information on domains, functions, products, ... in locus.",
-    # ~ )
-
-    # ~ parser.add_argument(
-    # ~ "--ripp_ann",
-    # ~ action="store_true",
-    # ~ help="TBA Add information on leader/core peptides, crosslinks, ... of RiPPs.",
-    # ~ )
-
-    # ~ parser.add_argument(
-    # ~ "--all",
-    # ~ action="store_true",
-    # ~ help="TBA Add full information suite at once",
-    # ~ )
-
     return parser
 
 
 def parse_arguments(VERSION: str, ROOT: Path) -> argparse.Namespace:
-    """Create program interface using argparse
+    """Create program interface using argparse.
 
     Args:
         `VERSION`: str indicating the version
@@ -82,7 +58,6 @@ def parse_arguments(VERSION: str, ROOT: Path) -> argparse.Namespace:
     Returns:
         `args` : Returns argparse object for further use
     """
-
     parser = add_parser_args(VERSION, ROOT)
     args = parser.parse_args()