+++ src/mibig_input_script/aux/verify_existence_entry.py
@@ -1,3 +1,11 @@
+"""
+Auxiliary function testing for presence of file to me modified.
+
+The user can specify an existing MIBiG entry to be modified. This function
+checks if the specified entry exists in one of the two directories and
+raises an error if directories are not found or if the file was not found.
+"""
+
 from pathlib import Path
 from typing import List
 
@@ -6,14 +14,14 @@ def verify_existence_entry(
     existing: List | None,
     ROOT: Path,
 ) -> Path | None:
-    """Test if MIBiG entry exists in folders and raise error if not
+    """Verify existence of directories/files.
 
     Parameters:
         `existing` : `List` OR `None`
         `ROOT` : `Path` object indicating "root" directory of script
 
     Returns:
-        `Path` OR `None` : a Path object pointing towards an MIBiG json file.
+        Path object pointing towards an MIBiG json file.
     """
     if existing is not None:
         mibig_entry = existing[0]
@@ -32,16 +40,18 @@ def verify_existence_entry(
         else:
             pass
 
-        if Path.exists(mibig_curr_dir_path.joinpath(mibig_entry).with_suffix(".json")):
-            return mibig_curr_dir_path.joinpath(mibig_entry).with_suffix(".json")
-        elif Path.exists(mibig_curr_dir_path.joinpath(mibig_entry)):
-            return mibig_curr_dir_path.joinpath(mibig_entry)
-        elif Path.exists(
-            mibig_next_dir_path.joinpath(mibig_entry).with_suffix(".json")
+        if Path.exists(
+            path := mibig_curr_dir_path.joinpath(mibig_entry).with_suffix(".json")
         ):
-            return mibig_next_dir_path.joinpath(mibig_entry).with_suffix(".json")
-        elif Path.exists(mibig_next_dir_path.joinpath(mibig_entry)):
-            return mibig_next_dir_path.joinpath(mibig_entry)
+            return path
+        elif Path.exists(path := mibig_curr_dir_path.joinpath(mibig_entry)):
+            return path
+        elif Path.exists(
+            path := mibig_next_dir_path.joinpath(mibig_entry).with_suffix(".json")
+        ):
+            return path
+        elif Path.exists(path := mibig_next_dir_path.joinpath(mibig_entry)):
+            return path
         else:
             message = (
                 f"Could not find the entry:\n"